pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "cthulhu follover"
authors = ["cthulhu follover <cthulhufollover@gmail.com>"]


[tool.taskipy.tasks]
manage = "python ./backend/manage.py"
runserver = "python ./backend/manage.py runserver"
migrate = "python ./backend/manage.py migrate"
makemigrations = "python ./backend/manage.py makemigrations"


[tool.black]
line-length = 79


[tool.isort]
multi_line_output=3
include_trailing_comma=true
use_parentheses=true
line_length=79


[tool.poetry.dependencies]
python = "^3.9, <3.10"
django = "^3.0.3"
django-environ = "^0.4.5"
psycopg2 = "^2.8.6"
django-rest-swagger = "^2.2.0"
django-extra-settings = "^0.1.3"
django-widget-tweaks = "^1.4.12"
Pillow = "^9.1.0"
django-model2puml = "^0.2.0"
django-debug-toolbar = "^3.4.0"



[tool.poetry.dev-dependencies]
pylint = "^2.4.4"
autopep8 = "^1.5"
pylint-django = "^2.0.13"
taskipy = "^1.1.3"
flake8 = "^3.7.9"
black = "^20.8b1"

